Javadoc for wrappers module added

+++ wrappers/src/main/java/cz/cuni/matfyz/collector/wrappers/neo4j/Neo4jDataCollector.java
@@ -13,17 +13,28 @@ import org.neo4j.driver.summary.ResultSummary;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Class responsible for collecting all statistical data from neo4j
+ */
 public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Result, String> {
 
     public Neo4jDataCollector(Neo4jConnection connection, DataModel model, String datasetName) {
         super(datasetName, model, connection);
     }
 
+    /**
+     * Method which saves page size used by databases storage engine
+     */
     private void _savePageSize() {
         _model.datasetData().setDataSetPageSize(Neo4jResources.DefaultSizes.PAGE_SIZE);
     }
 
-
+    /**
+     * Method responsible for parsing unit which is returned with cache size
+     * @param unit string representation of unit
+     * @return multiplier which corresponds to specified unit
+     * @throws ParseException when unit cannot be parsed
+     */
     private long _parseUnit(String unit) throws ParseException {
         if ("B".equals(unit))
             return 1;
@@ -41,7 +52,13 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
             throw new ParseException("Invalid unit for memory settings: " + unit);
     }
 
-    private int _parsePageCacheSize(String size) throws ParseException {
+    /**
+     * Method reponsible for parsing caching size
+     * @param size string which contains number and unit
+     * @return parsed number which is size in bytes
+     * @throws ParseException when input cannot be parsed
+     */
+    private long _parsePageCacheSize(String size) throws ParseException {
         StringBuilder number = new StringBuilder();
         StringBuilder unit = new StringBuilder();
         boolean isNumber = true;
@@ -57,8 +74,14 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
             }
         }
 
-        return (int) (Integer.parseInt(number.toString()) * _parseUnit(unit.toString()));
+        return (Integer.parseInt(number.toString()) * _parseUnit(unit.toString()));
     }
+
+    /**
+     * Method which saves the gahered cache size to data model
+     * @throws QueryExecutionException when QueryExecutionException occur in help query evaluation
+     * @throws ParseException when ParseException occur during unit parsing process
+     */
     private void _saveCacheSize() throws QueryExecutionException, ParseException {
         CachedResult result = _connection.executeQuery(Neo4jResources.getPageCacheSizeQuery());
         if (result.next()) {
@@ -68,6 +91,11 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
             }
         }
     }
+
+    /**
+     * Method responsible for saving all dataset sizes parameters to data model
+     * @throws QueryExecutionException when it is thrown from some of the help queries
+     */
     private void _saveDatasetSize() throws QueryExecutionException {
         long[] nodeTuple = _fetchNodePropertiesSize(Neo4jResources.getAllNodesQuery());
         long[] edgeTuple = _fetchEdgePropertiesSize(Neo4jResources.getAllRelationsQuery());
@@ -78,12 +106,26 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         ));
     }
 
+    /**
+     * Method which saves all dataset data
+     * @throws QueryExecutionException whe some of the help queries produced a problem
+     * @throws ParseException when there was some parsing issue with parsing cacheSize
+     */
     private void _saveDatasetData() throws QueryExecutionException, ParseException {
         _savePageSize();
         _saveDatasetSize();
         _saveCacheSize();
     }
 
+    // Save ColumnData
+
+    /**
+     * Mathod responsible for saving all column data
+     * @param label is label of node or edge from neo4j graoh
+     * @param property is field for which data we are interested in
+     * @param isNode boolean which indicates if we are parsing node property or edge property
+     * @throws QueryExecutionException when some of the help queries fail
+     */
     private void _saveSpecificColumnData(String label, String property, boolean isNode) throws QueryExecutionException {
         CachedResult result = isNode ? _connection.executeQuery(Neo4jResources.getNodePropertyTypeAndMandatoryQuery(label, property)) :
                 _connection.executeQuery(Neo4jResources.getEdgePropertyTypeAndMandatoryQuery(label, property));
@@ -96,6 +138,13 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
             _model.datasetData().setColumnMandatory(label, property, mandatory);
         }
     }
+
+    /**
+     * Method responsible for getting all properties for entity of some label
+     * @param query query which defines on what entities we are interested
+     * @return list of names for this properties
+     * @throws QueryExecutionException when help query fails
+     */
     private List<String> _getPropertyNames(String query) throws QueryExecutionException {
         List<String> properties = new ArrayList<>();
         CachedResult result = _connection.executeQuery(query);
@@ -107,6 +156,13 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         }
         return properties;
     }
+
+    /**
+     * Method responsible for saving all column data for entities of specific label
+     * @param tableName is label of entities
+     * @param isNode indicates if entity is node or edge
+     * @throws QueryExecutionException when some of the help queries fails
+     */
     private void _saveColumnData(String tableName, boolean isNode) throws QueryExecutionException {
         List<String> properties = isNode ? _getPropertyNames(Neo4jResources.getNodePropertiesForLabelQuery(tableName)) :
                 _getPropertyNames(Neo4jResources.getEdgePropertiesForLabelQuery(tableName));
@@ -116,7 +172,12 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         }
     }
 
-
+    /**
+     * Method which is responsible for calculating sizes of nodes of specific label
+     * @param fetchQuery defines on which nodes we are interested
+     * @return array of size 2, where first number is byteSize of nodes in the collection and second one is their count
+     * @throws QueryExecutionException when help query fails
+     */
     private long[] _fetchNodePropertiesSize(String fetchQuery) throws QueryExecutionException {
         ConsumedResult result = _connection.executeQueryAndConsume(fetchQuery);
         return new long[]{
@@ -124,12 +185,24 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         };
     }
 
+    /**
+     * Method which is responsible for calculating sizes of edges of specific label
+     * @param fetchQuery defines on which edges we are interested
+     * @return array of size 2, where first number is byteSize of edges in the collection and second one is their count
+     * @throws QueryExecutionException when help query fails
+     */
     private long[] _fetchEdgePropertiesSize(String fetchQuery) throws QueryExecutionException {
         ConsumedResult result = _connection.executeQueryAndConsume(fetchQuery);
         return new long[]{
                 result.getRowCount() * Neo4jResources.DefaultSizes.EDGE_SIZE + result.getByteSize(), result.getRowCount()
         };
     }
+
+    /**
+     * Method responsible for saving how much constraints were used on entities of specified label
+     * @param label speciufies entities we are interested in
+     * @throws QueryExecutionException when help query fails
+     */
     private void _saveTableConstraintCount(String label) throws QueryExecutionException {
         CachedResult result = _connection.executeQuery(Neo4jResources.getConstraintCountForLabelQuery(label));
         if (result.next()) {
@@ -138,6 +211,12 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         }
     }
 
+    /**
+     * Mathod responsible for saving sizes of entities of specific labels
+     * @param tableName is label of entities we are interested in
+     * @param isNode indicates if we are interested in nodes or edges
+     * @throws QueryExecutionException when some of the help queries fails
+     */
     private void _saveTableSizes(String tableName, boolean isNode) throws QueryExecutionException {
         long[] tuple = isNode ? _fetchNodePropertiesSize(Neo4jResources.getNodesOfSpecificLabelQuery(tableName)) : _fetchEdgePropertiesSize(Neo4jResources.getEdgesOfSpecificLabelQuery(tableName));
         long size = tuple[0];
@@ -149,6 +228,14 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         ));
         _model.datasetData().setTableRowCount(tableName, rowCount);
     }
+
+    /**
+     * Method which is responsible deciding if entity of specified label is node or edge
+     * @param label specified label
+     * @return true if label of entity is node label
+     * @throws QueryExecutionException when help query fails
+     * @throws DataCollectException when result of help query is empty
+     */
     private boolean _isNodeLabel(String label) throws QueryExecutionException, DataCollectException {
         CachedResult result = _connection.executeQuery(Neo4jResources.getIsNodeLabelQuery(label));
         if (result.next()) {
@@ -156,6 +243,12 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         }
         throw new DataCollectException("Invalid result of isNodeLabel query");
     }
+
+    /**
+     * Method which saves all data for entities used by query
+     * @throws QueryExecutionException when some of the help queries fails
+     * @throws DataCollectException when invalid label is used
+     */
     private void _saveTableData() throws QueryExecutionException, DataCollectException {
         for (String label : _model.getTableNames()) {
             boolean isNode = _isNodeLabel(label);
@@ -165,10 +258,21 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         }
     }
 
+    /**
+     * Method which gets labels over which is this index build
+     * @param indexNames specify index by its name
+     */
     private void _saveTableNameFor(String[] indexNames) {
         _model.datasetData().addTable(indexNames[1]);
     }
 
+    /**
+     * Method which is responsible for saving index sizes specified by index name over entities
+     * @param indexName identifier of index
+     * @param indexNames identifier split to array for usage of concrete tokens
+     * @param isNode indicates whether the entity is node or edge
+     * @throws QueryExecutionException when some of the help queries will fail
+     */
     private void _saveIndexSizes(String indexName, String[] indexNames, boolean isNode) throws QueryExecutionException{
         long[] index = isNode ? _fetchNodePropertiesSize(Neo4jResources.getNodeAndPropertyQuery(indexNames[1], indexNames[2])) :
                 _fetchEdgePropertiesSize(Neo4jResources.getEdgeAndPropertyQuery(indexNames[1], indexNames[2]));
@@ -182,6 +286,11 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         );
     }
 
+    /**
+     * Method for saving all data about used indexes
+     * @throws QueryExecutionException when some of the help queries fail
+     * @throws DataCollectException when some of the labels are invalid
+     */
     private void _saveIndexData() throws QueryExecutionException, DataCollectException {
         for (String inxName : _model.getIndexNames()) {
             String[] names = inxName.split(":");
@@ -192,6 +301,10 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
         }
     }
 
+    /**
+     * Method which is saving all statistics about result data to dataModel
+     * @param result result of main query
+     */
     private void _saveResultData(ConsumedResult result) {
         long size = result.getByteSize();
         _model.resultData().setByteSize(size);
@@ -208,6 +321,13 @@ public class Neo4jDataCollector extends AbstractDataCollector<ResultSummary, Res
             _model.resultData().setColumnByteSize(colName, Neo4jResources.DefaultSizes.getAvgColumnSizeByType(type));
         }
     }
+
+    /**
+     * Public method which triggers collecting of all statistical data and returns result as instance of DataModel
+     * @param result main query result
+     * @return instance of DataModel
+     * @throws DataCollectException when some of the help queries fails or invalid labels were used
+     */
     @Override
     public DataModel collectData(ConsumedResult result) throws DataCollectException {
         try {
